move secret loading tests to seperate class to make them independent of the example config
